feat: Group transfers by section and display them in the UI
