+++ app/src/main/java/com/infomaniak/swisstransfer/ui/previewparameter/TransferUiListPreviewParameterProvider.kt
@@ -26,48 +26,3 @@ import java.util.UUID
 class TransferUiListPreviewParameter : PreviewParameterProvider<List<TransferUi>> {
     override val values: Sequence<List<TransferUi>> = sequenceOf(transfersPreviewData)
 }
-
-val transfersPreviewData = listOf(
-    TransferUi(
-        uuid = UUID.randomUUID().toString(),
-        createdDateTimestamp = Date().time - 30L * 86_400_000L,
-        expirationDateTimestamp = Calendar.getInstance().apply {
-            time = Date()
-            set(Calendar.DATE, get(Calendar.DATE) + 3)
-        }.time.time,
-        sizeUploaded = 57_689_032L,
-        downloadLimit = 1,
-        downloadLeft = 1,
-        message = "Coucou c'est moi le message de description du transfert.",
-        password = "password",
-        files = filesPreviewData,
-    ),
-    TransferUi(
-        uuid = UUID.randomUUID().toString(),
-        createdDateTimestamp = Date().time - 5L * 86_400_000L,
-        expirationDateTimestamp = Calendar.getInstance().apply {
-            time = Date()
-            set(Calendar.DATE, get(Calendar.DATE) + 5)
-        }.time.time,
-        sizeUploaded = 89_723_143L,
-        downloadLimit = 20,
-        downloadLeft = 0,
-        message = null,
-        password = null,
-        files = filesPreviewData,
-    ),
-    TransferUi(
-        uuid = UUID.randomUUID().toString(),
-        createdDateTimestamp = Date().time - 0.5f.toLong() * 86_400_000L,
-        expirationDateTimestamp = Calendar.getInstance().apply {
-            time = Date()
-            set(Calendar.DATE, get(Calendar.DATE) - 4)
-        }.time.time,
-        sizeUploaded = 237_866_728L,
-        downloadLimit = 250,
-        downloadLeft = 123,
-        message = "3ème transfert. RAS.",
-        password = "my password",
-        files = filesPreviewData,
-    ),
-)
